HW1 Coach fix, change volatile long for int and use AtomicInteger for increment

+++ src/main/java/ru/javawebinar/topjava/storage/MealStorage.java
@@ -7,8 +7,12 @@ import java.util.List;
 public interface MealStorage {
 
     Meal add(Meal meal);
-    Meal get(long id);
+
+    Meal get(int id);
+
     Meal update(Meal meal);
-    void delete(long id);
+
+    void delete(int id);
+
     List<Meal> getAll();
 }
